// mobile/src/models/Local.ts
export enum LocalId {
    AEROPORTO = "AEROPORTO",
    BOULEVARD = "BOULEVARD",

    GAUCHO_VICENTE_FONTOURA = "GAUCHO_VICENTE_FONTOURA",
    SUPERMAGO_IPIRANGA = "SUPERMAGO_IPIRANGA",
    GAUCHO_ANTONIO_CARVALHO = "GAUCHO_ANTONIO_CARVALHO",
    SUPERMERCADO_FANTE = "SUPERMERCADO_FANTE",
    SUPERMAGO_PLANALTO = "SUPERMAGO_PLANALTO",
    SAMS_CLUB = "SAMS_CLUB",
    SUPERMAGO_BOA_VISTA = "SUPERMAGO_BOA_VISTA",
}

export interface Local {
    id: LocalId;
    nome: string;
}